Extract useTargetRecord into dedicated file
