+++ packages/twenty-front/src/modules/activities/timeline-activities/components/TimelineCard.tsx
@@ -4,10 +4,8 @@ import { CustomResolverFetchMoreLoader } from '@/activities/components/CustomRes
 import { SkeletonLoader } from '@/activities/components/SkeletonLoader';
 import { EventList } from '@/activities/timeline-activities/components/EventList';
 import { useTimelineActivities } from '@/activities/timeline-activities/hooks/useTimelineActivities';
-import {
-  useLayoutRenderingContext,
-  useTargetRecord,
-} from '@/ui/layout/contexts/LayoutRenderingContext';
+import { useLayoutRenderingContext } from '@/ui/layout/contexts/LayoutRenderingContext';
+import { useTargetRecord } from '@/ui/layout/contexts/useTargetRecord';
 import { useIsMobile } from '@/ui/utilities/responsive/hooks/useIsMobile';
 import {
   AnimatedPlaceholder,
