+++ packages/twenty-front/src/modules/dashboards/components/DashboardCard.tsx
@@ -2,7 +2,7 @@ import { DashboardContentRenderer } from '@/dashboards/components/DashboardConte
 import { type Dashboard } from '@/dashboards/components/types/Dashboard';
 import { CoreObjectNameSingular } from '@/object-metadata/types/CoreObjectNameSingular';
 import { useFindOneRecord } from '@/object-record/hooks/useFindOneRecord';
-import { useTargetRecord } from '@/ui/layout/contexts/LayoutRenderingContext';
+import { useTargetRecord } from '@/ui/layout/contexts/useTargetRecord';
 import { isDefined } from 'twenty-shared/utils';
 
 export const DashboardCard = () => {
